requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.10
openai>=1.10.0
paramiko>=3.4.0
pinecone-client>=3.1.0
beautifulsoup4>=4.12.3
requests>=2.31.0
tiktoken>=0.5.2
websockets>=12.0
python-multipart>=0.0.7
jinja2>=3.1.3
annotated-types>=0.6.0
llama-index-core>=0.10.11
llama-index-embeddings-openai>=0.1.6
llama-index-llms-openai>=0.1.6
llama-index-vector-stores-pinecone>=0.1.3
llama-index-postprocessor-sbert-rerank>=0.1.2
llama-parse>=0.3.0
sentence-transformers>=2.2.0
torch>=2.0.0

